refactor app into proper files
