+++ config.yaml
@@ -2,13 +2,15 @@ endpoints:
   - https://google.com
   - https://github.com
   - https://exam
-  
+
+#How often to ping in seconds
 ping_interval: 10
 
+#Alerting 
 alerting:
-  enabled: false
+  enabled: true
   consecutive_failures_threshold: 6
-  cooldown_duration: 300
+  cooldown_duration: 20
   pagerduty:
     enabled: false
     api_key: YOUR_PAGERDUTY_API_KEY
@@ -16,9 +18,9 @@ alerting:
     enabled: false
     api_key: YOUR_OPSGENIE_API_KEY
   slack:
-    enabled: false
-    webhook_url: YOUR_SLACK_WEBHOOK_URL
-    channel: alert-channel
+    enabled: true
+    webhook_url: https://hooks.slack.com/services/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
+    channel: alerts-channel
   telegram:
     enabled: false
     bot_token: YOUR_TELEGRAM_BOT_TOKEN
